Allow me to exclude from default_config
- Python isn't my native language so there are some glaring
  inefficiencies in this implementation.

- This allows you to add "exclude" to any of the integrations in the
config and will prevent that named integration from loading. There are
some caveats to how this is implemented that could cause some confusion
but I do not think it will in the real world usage. That's because if
you're using "default_config", your config is probably pretty simple and
terse.

In the config, we can now:

```yaml
default_config:
  ignore:
    - map
    - energy

```